Move modules back to System namespace
It started to feel a bit odd to have a `Data.Path` that uses graph
theory but is only for filepaths.  It seems this module namespace could
be useful for a less specialized path category library.

We're already heavily dependent on GHC so if you have to use other
packages that provide `System.Path` together with this package, it's not
that big of a deal to need `PackageImports` as well, or you can simply
import from `System.Path.Core` alone.

I rather like `System.Path` anyway.