Difference path experiment
With this we avoid the cost of concatenation, and get a Category in 7.6
without PolyKinds, and we can even compose with Prelude (.) and id
unless we newtype Hask (which we might want to do though so we can have
sensible instances like IsString or for that matter Resolve).

Downsides included added complexity and the need to apply Nil to get an
actual Path that we can manipulate.  This might not matter much for most
users though if we newtype Hask so we can Resolve a "dpath" directly,
etc.