Rename Vertex and Edge to File and Link
Borrowing terminology from POSIX filesystems rather than graph theory.
The old 'File' vertex is renamed to 'Regular' both to avoid the need to
disambiguate between the promoted and normal 'File' type and because
when we go along with "everything is a file" the term "file" alone is
too general for regular files.

Closes #1